pcbnew: Allow tuning length to be longer than INT_MAX
The INT_MAX limit for most elements makes sense only for single-segment,
straight line elements.  For elements that accumulate lengths, we should
utilize the long long int (64 bits) to allow for greater lengths.

Fixes: lp:1842367
* https://bugs.launchpad.net/kicad/+bug/1842367

+++ common/geometry/shape_line_chain.cpp
@@ -100,9 +100,9 @@ const SHAPE_LINE_CHAIN SHAPE_LINE_CHAIN::Reverse() const
 }
 
 
-int SHAPE_LINE_CHAIN::Length() const
+long long int SHAPE_LINE_CHAIN::Length() const
 {
-    int l = 0;
+    long long int l = 0;
 
     for( int i = 0; i < SegmentCount(); i++ )
         l += CSegment( i ).Length();
